requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-Caching==2.1.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.11
openai==1.3.7
schedule==1.2.0
gunicorn==21.2.0
python-socketio==5.9.0
eventlet==0.33.3
tiktoken==0.5.2
psycopg2-binary==2.9.9